src/python-common/requirements.txt
six
pytest >=2.1.3,<5; python_version < '3.5'
mock; python_version < '3.3'
mypy==0.782; python_version >= '3'
pytest-mypy; python_version >= '3'
pytest >= 2.1.3; python_version >= '3'
pyyaml
kubernetes